social-content-automation/requirements.txt
Pillow>=10.0.0
python-pptx>=0.6.21
pyyaml>=6.0
matplotlib>=3.7.0
pandas>=2.0.0
imageio>=2.25.0
moviepy>=1.0.3
beautifulsoup4>=4.12.0
requests>=2.31.0
